+++ tests/pg-promise/index.test.mjs
@@ -1,5 +1,6 @@
 import pgpromise from 'pg-promise';
 import pgvector from 'pgvector/pg-promise';
+import { SparseVector } from 'pgvector';
 
 test('example', async () => {
   const initOptions = {
@@ -12,21 +13,30 @@ test('example', async () => {
 
   await db.none('CREATE EXTENSION IF NOT EXISTS vector');
   await db.none('DROP TABLE IF EXISTS pg_promise_items');
-  await db.none('CREATE TABLE pg_promise_items (id serial PRIMARY KEY, embedding vector(3))');
+  await db.none('CREATE TABLE pg_promise_items (id serial PRIMARY KEY, embedding vector(3), half_embedding halfvec(3), binary_embedding bit(3), sparse_embedding sparsevec(3))');
 
   const params = [
-    pgvector.toSql([1, 1, 1]),
-    pgvector.toSql([2, 2, 2]),
-    pgvector.toSql([1, 1, 2]),
-    null
+    pgvector.toSql([1, 1, 1]), pgvector.toSql([1, 1, 1]), '000', SparseVector.fromDense([1, 1, 1]).toSql(),
+    pgvector.toSql([2, 2, 2]), pgvector.toSql([2, 2, 2]), '101', SparseVector.fromDense([2, 2, 2]).toSql(),
+    pgvector.toSql([1, 1, 2]), pgvector.toSql([1, 1, 2]), '111', SparseVector.fromDense([1, 1, 2]).toSql(),
+    null, null, null, null
   ];
-  await db.none('INSERT INTO pg_promise_items (embedding) VALUES ($1), ($2), ($3), ($4)', params);
+  await db.none('INSERT INTO pg_promise_items (embedding, half_embedding, binary_embedding, sparse_embedding) VALUES ($1, $2, $3, $4), ($5, $6, $7, $8), ($9, $10, $11, $12), ($13, $14, $15, $16)', params);
 
   const items = await db.any('SELECT * FROM pg_promise_items ORDER BY embedding <-> $1 LIMIT 5', [pgvector.toSql([1, 1, 1])]);
   expect(items.map(v => v.id)).toStrictEqual([1, 3, 2, 4]);
   expect(items[0].embedding).toStrictEqual([1, 1, 1]);
   expect(items[1].embedding).toStrictEqual([1, 1, 2]);
   expect(items[2].embedding).toStrictEqual([2, 2, 2]);
+  expect(items[0].half_embedding).toStrictEqual([1, 1, 1]);
+  expect(items[1].half_embedding).toStrictEqual([1, 1, 2]);
+  expect(items[2].half_embedding).toStrictEqual([2, 2, 2]);
+  expect(items[0].binary_embedding).toStrictEqual('000');
+  expect(items[1].binary_embedding).toStrictEqual('111');
+  expect(items[2].binary_embedding).toStrictEqual('101');
+  expect(items[0].sparse_embedding.toArray()).toStrictEqual([1, 1, 1]);
+  expect(items[1].sparse_embedding.toArray()).toStrictEqual([1, 1, 2]);
+  expect(items[2].sparse_embedding.toArray()).toStrictEqual([2, 2, 2]);
 
   await db.none('CREATE INDEX ON pg_promise_items USING hnsw (embedding vector_l2_ops)');